Created ClientPortalDialog component
